pyproject.toml
[project]
name = "sea-mt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.13.0",
    "aqlm[cpu,gpu]>=1.1.7",
    "bert-score>=0.3.13",
    "bs4>=0.0.2",
    "deepspeed>=0.18.8",
    "evaluate>=0.4.6",
    "flash-linear-attention>=0.4.2",
    "flashinfer-jit-cache>=0.6.7",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "jupytext>=1.19.1",
    "liger-kernel>=0.7.0",
    "malaya>=5.1.1",
    "matplotlib>=3.10.8",
    "nltk>=3.9.4",
    "openrlhf[vllm]==0.9.8",
    "peft>=0.18.1",
    "rouge-score>=0.1.2",
    "sacrebleu>=2.6.0",
    "shap>=0.51.0",
    "torch>=2.10.0",
    "torch-optimi>=0.3.3",
    "torchao==0.16.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "trl>=0.29.1",
    "unbabel-comet>=2.2.7",
    "vllm>=0.18.0",
    "wandb>=0.25.1",
    "xielu>=0.1",
]

[[tool.uv.index]]
url = "https://flashinfer.ai/whl/cu130"

[[tool.uv.index]]
name = "pytorch-130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-130"},
]
torchvision = [
  { index = "pytorch-130"},
]
torchaudio = [
  { index = "pytorch-130"},
]
torchao = [
  { index = "pytorch-130"},
]
xielu = { git = "https://github.com/rubber-duck-debug/xielu.git" }

[tool.uv]
override-dependencies = ["numpy==2.2.6", "protobuf==6.33.5", "transformers>=5.0.0", "huggingface-hub>=1.0.0"]
exclude-dependencies = ['pynvml']

[tool.uv.extra-build-dependencies]
flash-attn = ["torch"]
xielu = ["torch"]
